sensu-go-check: Migrate to Sensu Go API
This commit contains the new version of the work-in-progress
Ansible module 'sensu-go-check'. It's been completely re-written from
the ground up to no longer depend on 'sensuctl' and instead directly
integrate with the Sensu Go API v2.

Please note the module is still in 'preview' status and may not be
perfect. This change completely breaks the the behaviour from the
previous version.

Signed-off-by: Jared Ledvina <jared@techsmix.net>